slide_puzzle/requirements.txt
pytest
pygame
Pillow>=5.0.0
